+++ src/occupancy_data/empty_occupancy.rs
